changed floats to doubles in Bod2D class

+++ src/map/bod/Bod.java
@@ -2,13 +2,13 @@ package map.bod;
 
 public interface Bod
 {
-    float getX();
-    float getY();
-    float[] getXY();
+    double getX();
+    double getY();
+    double[] getXY();
     void setX(float _x);
     void setY(float _y);
     void setXY(float ...suradnice);
-    float getDistance(Bod other);
+    double getDistance(Bod other);
     Bod addBod(Bod other);
     Bod subtractBod(Bod other);
     boolean equals(Bod other);
